refactor: lib 내부 폴더구조 세분화 - util / hooks / animations
