+++ src/shared/lib/animations/animations.ts
