+++ src/shared/lib/utils/getProgressDegree.ts
