+++ src/shared/lib/utils/handleImageError.ts
@@ -1,4 +1,4 @@
-import fallbackIcon from '../assets/icons/ic_fallback.svg';
+import fallbackIcon from '../../assets/icons/ic_fallback.svg';
 
 export const handleImageError = (e: React.SyntheticEvent<HTMLImageElement>) => {
   e.currentTarget.src = fallbackIcon;
